// Views/Settings/Enums.ts
export enum MessageType {
    CATALOGUE = 'catalogue',
    PRODUCT = 'product'
}

export enum Selector {
    catalogueSetting = '#catalogue-setting',
    blockPanel_Catalogue = '#block-panel-catalogue',
    blockPanel_Catalogue_RemoveButton_Last = '#block-panel-catalogue .remove-btn:last',
    blockPanel_Catalogue_Input_Last = '#block-panel-catalogue .custom-block:last',
    componentSelect_Catalogue = '#component-select-catalogue',
    addComponentButton_Catalogue = '#component-btn-catalogue',
    customTextInput_Catalogue = '#custom-input-catalogue',
    addCustomTextButton_Catalogue = '#custom-btn-catalogue',
    restoreDefaultButton_Catalogue = '#restore-default-btn-catalogue',
    saveTemplateButton_Catalogue = '#save-template-btn-catalogue',
    previewPanel_Catalogue = '#preview-panel-catalogue',
    blockPanel_Product = '#block-panel-product',
    blockPanel_Product_RemoveButton_Last = '#block-panel-product .remove-btn:last',
    blockPanel_Product_Input_Last = '#block-panel-product .custom-block:last',
    componentSelect_Product = '#component-select-product',
    addComponentButton_Product = '#component-btn-product',
    customTextInput_Product = '#custom-input-product',
    addCustomTextButton_Product = '#custom-btn-product',
    restoreDefaultButton_Product = '#restore-default-btn-product',
    saveTemplateButton_Product = '#save-template-btn-product',
    previewPanel_Product = '#preview-panel-product',
    loadIndicator_Top = '#load-indicator-top',
    keywordInput_Product = '#keyword-input-product'
}